webservices_rpi_arduino_comm/requirements.txt
alabaster==0.7.10
AntClient==0.4.0
apipkg==1.5
asn1crypto==1.0.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.7.0
backports.functools-lru-cache==1.5
bcrypt==3.1.7
blessings==1.6.1
boolean==1.1.0
boto==2.3.0
boto3==1.9.12
botocore==1.12.12
bugsnag==3.1.2
cffi==1.11.5
click==6.7
constants==0.6.0
coverage==4.5.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
docutils==0.12
ecdsa==0.13
enum34==1.1.6
execnet==1.5.0
flake8==2.4.0
Flask==1.1.1
Flask-Cors==3.0.2
funcsigs==1.0.2
futures==3.2.0
httmock==1.2.5
imagesize==1.0.0
inquirer==2.2.0
ipaddress==1.0.22
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10.3
jmespath==0.9.3
jsonpath-rw==1.4.0
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mccabe==0.3.1
mock==2.0.0
more-itertools==4.3.0
msgpack-python==0.1.12
numpy==1.16.5
paramiko==2.6.0
pathlib2==2.3.2
pbr==5.4.3
pep8==1.5.7
pjsua==2.9
pkgconfig==1.4.0
pluggy==0.7.1
ply==3.11
prompt-toolkit==1.0.15
psutil==4.3.0
psycopg2==2.7.5
py==1.5.3
pycparser==2.18
pycrypto==2.6.1
pyflakes==0.8.1
Pygments==2.2.0
PyNaCl==1.3.0
pyparsing==2.4.2
pyserial==3.4
pystache==0.5.4
pytest==2.9.2
pytest-cov==2.3.0
pytest-mock==1.5.0
pytest-xdist==1.14
python-dateutil==2.7.3
pytz==2019.2
readchar==0.7
requests==2.10.0
robotframework==3.0
robotframework-ride==1.5.2.1
robotframework-selenium2library==1.7.4
robotframework-sshlibrary==2.1.2
robotremoteserver==1.0.1
s3transfer==0.1.13
scandir==1.9.0
selenium==3.12.0
six==1.11.0
smpplib==0.1
snowballstemmer==1.2.1
Sphinx==1.4.6
sphinxcontrib-robotdoc==0.8.0
sphinxcontrib-robotframework==0.5.1
ssh==1.8.0
subprocess32==3.5.4
tabulate==0.8.2
termcolor==1.1.0
tox==3.2.1
traitlets==4.2.2
urllib3==1.23
validators==0.12.1
virtualenv==16.0.0
Wave==0.0.2
wcwidth==0.1.7
WebOb==1.8.5
Werkzeug==0.16.0
xmlsec==1.3.3
yasm==0.0
